+++ client/src/entrypoints/admin/telepath/widgets.js
@@ -128,26 +128,6 @@ class Select extends Widget {
 }
 window.telepath.register('wagtail.widgets.Select', Select);
 
-class PageChooserFactory {
-  constructor(html, idPattern, config) {
-    this.html = html;
-    this.idPattern = idPattern;
-    this.config = config;
-  }
-
-  render(placeholder, name, id, initialState) {
-    const html = this.html.replace(/__NAME__/g, name).replace(/__ID__/g, id);
-    // eslint-disable-next-line no-param-reassign
-    placeholder.outerHTML = html;
-    /* the PageChooser object also serves as the JS widget representation */
-    // eslint-disable-next-line no-undef
-    const chooser = new PageChooser(id, null, this.config);
-    chooser.setState(initialState);
-    return chooser;
-  }
-}
-window.telepath.register('wagtail.widgets.PageChooser', PageChooserFactory);
-
 class AdminAutoHeightTextInput extends Widget {
   render(placeholder, name, id, initialState, parentCapabilities) {
     const boundWidget = super.render(
